requirements.txt
langchain
langchain-community
langchain-openai
chromadb
openai
python-dotenv
gradio
sentence-transformers
streamlit